everything works!!

+++ python/xpctl/backends/sql/backend.py
@@ -151,14 +151,15 @@ class SQLRepo(ExperimentRepo):
         
     def get_model_location(self, task, eid):
         session = self.Session()
-        exp = session.query(SqlExperiment).filter(SqlExperiment.task == task).filter(SqlExperiment.eid == eid)
-        if exp is None or exp.scalar() is None:
-            return Error(message='no model location for experiment id [{}] in [{}] database'.format(eid, task))
-        return exp.checkpoint
+        exp = session.query(SqlExperiment).get(eid)
+        return Success(exp.checkpoint)
 
-    def update_prop(self, task, eid, prop, value):  # TODO: not using task
+    def update_prop(self, task, eid, prop, value):
         try:
             session = self.Session()
+            exp = session.query(SqlExperiment).filter(SqlExperiment.task == task).filter(SqlExperiment.eid == eid)
+            if exp is None or exp.scalar() is None:
+                return Error(message='no experiment with id [{}] for task [{}]'.format(eid, task))
             # there must be a better way of getting a column value through a column name
             prev_value = getattr(session.query(SqlExperiment).get(eid), prop)
             session.query(SqlExperiment).filter(SqlExperiment.eid == eid).update({prop: value})
@@ -169,18 +170,20 @@ class SQLRepo(ExperimentRepo):
         except sql.exc.SQLAlchemyError as e:
             return Error(message=str(e))
 
-    def remove_experiment(self, task, eid):  # TODO: not using task
+    def remove_experiment(self, task, eid):
         try:
             session = self.Session()
-            exp = session.query(SqlExperiment).get(eid)
-            if exp is None:
-                return Error(message='delete failed: experiment {} not found in {} database'.format(eid, task))
-            model_loc = self.get_model_location(task, eid)
-            if model_loc is not None and type(model_loc) is not Error and os.path.exists(model_loc):
+            exp = session.query(SqlExperiment).filter(SqlExperiment.task == task).filter(SqlExperiment.eid == eid)
+            if exp is None or exp.scalar() is None:
+                return Error(message='no experiment with id [{}] for task [{}]'.format(eid, task))
+            model_loc_response = self.get_model_location(task, eid)
+            model_loc = model_loc_response.message
+            if model_loc is not None and type(model_loc_response) is not Error and os.path.exists(model_loc):
                 try:
                     os.remove(model_loc)
                 except IOError:
                     return Error(message='model {} exists on host but could not be removed'.format(model_loc))
+            exp = session.query(SqlExperiment).get(eid)
             session.delete(exp)
             session.commit()
             try:
@@ -207,7 +210,7 @@ class SQLRepo(ExperimentRepo):
         reduction_dim = reduction_dim if reduction_dim is not None else 'sha1'
         hits = session.query(SqlExperiment).filter(getattr(SqlExperiment, prop) == value)\
             .filter(SqlExperiment.task == task)
-        if hits is None:
+        if hits is None or not hits.first():
             return Error(message='no information available for [{}]: [{}] in task database [{}]'
                          .format(prop, value, task))
         data_experiments = []
@@ -217,8 +220,19 @@ class SQLRepo(ExperimentRepo):
                 return data_experiment
             else:
                 data_experiments.append(data_experiment)
-        return aggregate_sql_results(data_experiments, reduction_dim, event_type, numexp_reduction_dim)
- 
+        experiment_aggregate_set = aggregate_sql_results(data_experiments, reduction_dim, event_type,
+                                                         numexp_reduction_dim)
+        if sort is None or (type(sort) == str and sort == 'None'):
+            return experiment_aggregate_set
+        else:
+            if event_type == 'test_events':
+                if sort in METRICS_SORT_ASCENDING:
+                    return experiment_aggregate_set.sort(sort, reverse=False)
+                else:
+                    return experiment_aggregate_set.sort(sort)
+            else:
+                return Error(message='experiments can only be sorted when event_type=test_events')
+         
     def list_results(self, task, prop, value, user, metric, sort, event_type):
         session = self.Session()
         sql_experiments = []
@@ -228,6 +242,8 @@ class SQLRepo(ExperimentRepo):
             filter(SqlExperiment.task == task)
         if users:
             hits = hits.filter(SqlExperiment.username.in_(users)).all()
+        if hits is None or not hits.first():
+            return Error('No results in {} database for {} = {}'.format(task, prop, value))
         for exp in hits:
             sql_experiments.append(sql_result_to_data_experiment(exp, event_type, metrics))
         experiment_set = get_data_experiment_set(sql_experiments)
